pyproject.toml
[project]
name = "heatmap-of-green-history"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.16",
    "aiosignal==1.3.2",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "botocore==1.37.33",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "frozenlist==1.5.0",
    "gql==3.5.2",
    "graphql-core==3.2.4",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "idna==3.10",
    "jmespath==1.0.1",
    "multidict==6.4.3",
    "numpy==2.2.4",
    "pandas==2.2.3",
    "propcache==0.3.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "websockets==11.0.3",
    "yarl==1.19.0",
]
